[feat][fix] use guideline and bias to define the place of bottom sheet
With guideline (and 'constrainedHeight') to make sure bottom sheet will
never be too high.

Figured out that the 'parent' of an included layout is the layout
width / height defined in the include manifesto.  It is better to set
layout width / height to 'match_parent' in the included layout.  This
could also be a potential bug.