[feature] add an option to bypass the exposure time limit
Each device has their upper limit of exposure time declared in
`cameraCharacteristics`.  The documentation said any value greater than
then maximum will be cut down to maximum, but it actually did not.  As
the exposure time exceeding the upper limit, it just become more
unstable; if it did not crash, it can indeed capture the shoot.